rating in double format

+++ app/src/main/java/com/example/complamap/model/User.kt
@@ -6,6 +6,6 @@ data class User(
     val username: String? = null,
     val email: String? = null,
     val profilePic: String? = null,
-    val rating: Long? = null,
+    val rating: Double? = null,
     val subs: List<DocumentReference>? = null,
 )
